Full API

Autogenerated API list

SymplecticMatrices.BlochMessiah — Type

BlochMessiah <: Factorization

Matrix factorization type of the Bloch-Messiah/Euler decomposition of a symplectic matrix S. This is the return type of blochmessiah(_), the corresponding matrix factorization function.

If F::BlochMessiah is the factorization object, O, values and Qcan be obtained viaF.O,F.values, andF.Q`.

Iterating the decomposition produces the components O, values, and Q, in that order.

Examples

julia> S = Symplectic(BlockForm(1), [1. 1.; 0. 1.])
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
 1.0  1.0
 0.0  1.0

julia> issymplectic(S)
true

julia> F = blochmessiah(S)
BlochMessiah{Float64, Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}, Vector{Float64}}
O factor:
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
 0.850651  -0.525731
 0.525731   0.850651
values:
1-element Vector{Float64}:
 1.618033988749895
Q factor:
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
  0.525731  0.850651
 -0.850651  0.525731

julia> isapprox(S, F.O * Diagonal(vcat(F.values, F.values .^ (-1))) * F.Q, atol = 1e-10)
true

julia> issymplectic(F.O, atol = 1e-10) && issymplectic(F.Q, atol = 1e-10)
true

julia> O, values, Q = F; # destructuring via iteration

julia> O == F.O && values == F.values && Q == F.Q
true

SymplecticMatrices.Polar — Type

Polar <: Factorization

Matrix factorization type of the polar decomposition of a symplectic matrix S. This is the return type of polar(_), the corresponding matrix factorization function.

If F::Polar is the factorization object, O and P can be obtained via F.O and F.P, such that S = O * P.

Iterating the decomposition produces the components O and P.

Examples

julia> S = [1. 1.; 0. 1.]
2×2 Matrix{Float64}:
 1.0  1.0
 0.0  1.0

julia> issymplectic(BlockForm(1), S)
true

julia> F = polar(S)
Polar{Float64, Matrix{Float64}, Matrix{Float64}}
O factor:
2×2 Matrix{Float64}:
  0.894427  0.447214
 -0.447214  0.894427
P factor:
2×2 Matrix{Float64}:
 0.894427  0.447214
 0.447214  1.34164

julia> isapprox(F.O * F.P, S)
true

julia> O, P = F; # destructuring via iteration

julia> O == F.O && P == F.P
true

SymplecticMatrices.Symplectic — Method

Symplectic(J::SymplecticForm, S::AbstractMatrix) <: AbstractMatrix

Construct a wrapper of a symplectic matrix S with its corresponding symplectic basis defined by the symplectic form J.

SymplecticMatrices.SymplecticGauss — Type

SymplecticGauss <: AbstractMatrix

Matrix representation type of a symplectic Gauss transformation on a symplectic basis. This is the return type of gauss(_), the corresponding transformation matrix function.

If G::SymplecticGauss is the transformation object, the mode k and parameters c and d can be obtained via G.k, G.c, and G.d.

Examples

julia> G = gauss(PairForm(3), 2, 2.0, 3.0)
6×6 SymplecticGauss{PairForm{Int64}, Int64, Float64}:
 2.0   ⋅    ⋅   3.0   ⋅    ⋅ 
  ⋅   0.5   ⋅    ⋅    ⋅    ⋅ 
  ⋅   3.0  2.0   ⋅    ⋅    ⋅ 
  ⋅    ⋅    ⋅   0.5   ⋅    ⋅ 
  ⋅    ⋅    ⋅    ⋅   1.0   ⋅ 
  ⋅    ⋅    ⋅    ⋅    ⋅   1.0

julia> isapprox(G * inv(G), I)
true

SymplecticMatrices.SymplecticGivens — Type

SymplecticGivens <: AbstractMatrix

Matrix representation type of a symplectic Givens rotation on a symplectic basis. This is the return type of givens(_), the corresponding rotation matrix function.

If G::SymplecticGivens is the rotation object, the mode k, cosine c, and sine s can be obtained via G.k, G.c, and G.s, such that c^2 + s^2 = 1.

Examples

julia> G = givens(BlockForm(3), 2, pi/4)
SymplecticGivens{BlockForm{Int64}, Int64, Float64}:
 1.0    ⋅                   ⋅    ⋅    ⋅                   ⋅ 
  ⋅    0.7071067811865476   ⋅    ⋅   0.7071067811865475   ⋅ 
  ⋅     ⋅                  1.0   ⋅    ⋅                   ⋅ 
  ⋅     ⋅                   ⋅   1.0   ⋅                   ⋅ 
  ⋅   -0.7071067811865475   ⋅    ⋅   0.7071067811865476   ⋅ 
  ⋅     ⋅                   ⋅    ⋅    ⋅                  1.0

julia> k, c, s = G.k, G.c, G.s;

julia> c^2 + s^2 ≈ 1.0
true

SymplecticMatrices.SymplecticHouseholder — Type

SymplecticHouseholder <: AbstractMatrix

Matrix representation type of a symplectic Householder reflection on a symplectic basis. This is the return type of householder(_), the corresponding reflection matrix function.

If H::SymplecticHouseholder is the reflection object, the mode k and projection matrix P can be obtained via H.k and H.P.

Examples

julia> H = householder(BlockForm(3), 2, [0.23, 0.45])
6×6 SymplecticHouseholder{BlockForm{Int64}, Int64, Matrix{Float64}}:
 1.0    ⋅          ⋅         ⋅     ⋅          ⋅ 
  ⋅    0.585748  -0.810493   ⋅     ⋅          ⋅ 
  ⋅   -0.810493  -0.585748   ⋅     ⋅          ⋅ 
  ⋅     ⋅          ⋅        1.0    ⋅          ⋅ 
  ⋅     ⋅          ⋅         ⋅    0.585748  -0.810493
  ⋅     ⋅          ⋅         ⋅   -0.810493  -0.585748

julia> isapprox(H * H, I)
true

SymplecticMatrices.Williamson — Type

Williamson <: Factorization

Matrix factorization type of the williamson decomposition of a positive-definite matrix V. This is the return type of williamson(_), the corresponding matrix factorization function.

If F::Williamson is the factorization object, S and spectrum can be obtained via F.S and F.spectrum.

Iterating the decomposition produces the components S and spectrum.

Examples

julia> V = [7. 2.; 2. 1.]
2×2 Matrix{Float64}:
 7.0  2.0
 2.0  1.0

julia> isposdef(V)
true

julia> F = williamson(BlockForm(1), V)
Williamson{Float64, Matrix{Float64}, Vector{Float64}}
S factor:
2×2 Matrix{Float64}:
 0.448828  -1.95959
 0.61311   -0.448828
symplectic spectrum:
1-element Vector{Float64}:
 1.7320508075688772

julia> isapprox(F.S * V * F.S', Diagonal(repeat(F.spectrum, 2)))
true

julia> S, spectrum = F; # destructuring via iteration

julia> S == F.S && spectrum == F.spectrum
true

LinearAlgebra.givens — Method

givens(form::SymplecticForm, k::Int, c::Number, s::Number) -> SymplecticGivens

Compute a symplectic Givens rotation of mode k and return a SymplecticGivens object.

The parameters c and s represent the cosine and sine of the rotation angle, satisfying c^2 + s^2 = 1. The transformation G performs a rotation in the (k, n+k) plane for BlockForm or the (2k-1, 2k) plane for PairForm.

SymplecticMatrices.blochmessiah — Method

blochmessiah(form::SymplecticForm, S::AbstractMatrix) -> BlochMessiah
blochmessiah(S::Symplectic) -> BlochMessiah

Compute the Bloch-Messiah/Euler decomposition of a symplectic matrix S and return a BlockMessiah object.

The orthogonal symplectic matrices O and Q as well as the singular values values can be obtained via F.O, F.Q, and F.values, respectively.

Iterating the decomposition produces the components O, values, and Q, in that order.

Examples

julia> S = Symplectic(BlockForm(1), [1. 1.; 0. 1.])
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
 1.0  1.0
 0.0  1.0

julia> issymplectic(S)
true

julia> F = blochmessiah(S)
BlochMessiah{Float64, Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}, Vector{Float64}}
O factor:
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
 0.850651  -0.525731
 0.525731   0.850651
values:
1-element Vector{Float64}:
 1.618033988749895
Q factor:
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
  0.525731  0.850651
 -0.850651  0.525731

julia> isapprox(S, F.O * Diagonal(vcat(F.values, F.values .^ (-1))) * F.Q, atol = 1e-10)
true

julia> issymplectic(F.O, atol = 1e-10) && issymplectic(F.Q, atol = 1e-10)
true

julia> O, values, Q = F; # destructuring via iteration

julia> O == F.O && values == F.values && Q == F.Q
true

SymplecticMatrices.gauss — Method

gauss(form::SymplecticForm, k::Int, c::Number, d::Number) -> SymplecticGauss

Compute a symplectic Gauss transformation of mode k and return a SymplecticGauss object.

The parameters c and d define the scaling and shear of the transformation.

SymplecticMatrices.householder — Method

householder(form::SymplecticForm, k::Int, v::AbstractVector) -> SymplecticHouseholder

Compute a symplectic Householder transformation of mode k and return a SymplecticHouseholder object.

The transformation uses a Householder reflector P = I - 2vvᵀ / vᵀv to act simultaneously on the leading and trailing blocks of the symplectic basis. Starting from index k, it reflects the corresponding sub-vectors in both halves of the space.

SymplecticMatrices.issymplectic — Method

issymplectic(form::Symplecticform, x::AbstractMatrix; atol=0.0, rtol=atol)
issymplectic(x::Symplectic; atol=0.0, rtol=atol)

Return whether or not an input matrix is symplectic. Keyword arguments atol and rtol can be called to determine absolute and relative tolerances of the check, respectively.

SymplecticMatrices.polar — Method

polar(S::AbstractMatrix) -> Polar
polar(S::Symplectic) -> Polar

Compute the polar decomposition of a symplectic matrix S and return a Polar object.

O and P can be obtained from the factorization F via F.O and F.P, such that S = O * P. For the symplectic polar decomposition case, O is an orthogonal symplectic matrix and P is a positive-definite symmetric symplectic matrix.

Iterating the decomposition produces the components O and P.

Examples

julia> S = [1. 1.; 0. 1.]
2×2 Matrix{Float64}:
 1.0  1.0
 0.0  1.0

julia> issymplectic(BlockForm(1), S)
true

julia> F = polar(S)
Polar{Float64, Matrix{Float64}, Matrix{Float64}}
O factor:
2×2 Matrix{Float64}:
  0.894427  0.447214
 -0.447214  0.894427
P factor:
2×2 Matrix{Float64}:
 0.894427  0.447214
 0.447214  1.34164

julia> isapprox(F.O * F.P, S)
true

julia> O, P = F; # destructuring via iteration

julia> O == F.O && P == F.P
true

SymplecticMatrices.randsymplectic — Method

randsymplectic(form::SymplecticForm)
randsymplectic(::Symplectic, form::SymplecticForm)

Calculate a random symplectic matrix in symplectic representation defined by basis.

SymplecticMatrices.williamson — Method

williamson(form::SymplecticForm, V::AbstractMatrix) -> Williamson
williamson(::Symplectic, form::SymplecticForm, V::AbstractMatrix) -> Williamson

Compute the williamson decomposition of a positive-definite matrix V and return a Williamson object.

A symplectic matrix S and symplectic spectrum spectrum can be obtained via F.S and F.spectrum.

Iterating the decomposition produces the components S and spectrum.

Examples

julia> V = [7. 2.; 2. 1.]
2×2 Matrix{Float64}:
 7.0  2.0
 2.0  1.0

julia> isposdef(V)
true

julia> F = williamson(BlockForm(1), V)
Williamson{Float64, Matrix{Float64}, Vector{Float64}}
S factor:
2×2 Matrix{Float64}:
 0.448828  -1.95959
 0.61311   -0.448828
symplectic spectrum:
1-element Vector{Float64}:
 1.7320508075688772

julia> isapprox(F.S * V * F.S', Diagonal(repeat(F.spectrum, 2)))
true

julia> S, spectrum = F; # destructuring via iteration

julia> S == F.S && spectrum == F.spectrum
true