Full API

Autogenerated API list

SymplecticMatrices.BlochMessiah — Type

BlochMessiah <: Factorization

Matrix factorization type of the Bloch-Messiah/Euler decomposition of a symplectic matrix S. This is the return type of blochmessiah(_), the corresponding matrix factorization function.

If F::BlochMessiah is the factorization object, O, values and Qcan be obtained viaF.O,F.values, andF.Q`.

Iterating the decomposition produces the components O, values, and Q, in that order.

Examples

julia> S = Symplectic(BlockForm(1), [1. 1.; 0. 1.])
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
 1.0  1.0
 0.0  1.0

julia> issymplectic(S)
true

julia> F = blochmessiah(S)
BlochMessiah{Float64, Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}, Vector{Float64}}
O factor:
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
 0.850651  -0.525731
 0.525731   0.850651
values:
1-element Vector{Float64}:
 1.618033988749895
Q factor:
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
  0.525731  0.850651
 -0.850651  0.525731

julia> isapprox(S, F.O * Diagonal(vcat(F.values, F.values .^ (-1))) * F.Q, atol = 1e-10)
true

julia> issymplectic(F.O, atol = 1e-10) && issymplectic(F.Q, atol = 1e-10)
true

julia> O, values, Q = F; # destructuring via iteration

julia> O == F.O && values == F.values && Q == F.Q
true

SymplecticMatrices.Polar — Type

Polar <: Factorization

Matrix factorization type of the polar decomposition of a symplectic matrix S. This is the return type of polar(_), the corresponding matrix factorization function.

If F::Polar is the factorization object, O and P can be obtained via F.O and F.P, such that S = O * P.

Iterating the decomposition produces the components O and P.

Examples

julia> S = [1. 1.; 0. 1.]
2×2 Matrix{Float64}:
 1.0  1.0
 0.0  1.0

julia> issymplectic(BlockForm(1), S)
true

julia> F = polar(S)
Polar{Float64, Matrix{Float64}, Matrix{Float64}}
O factor:
2×2 Matrix{Float64}:
  0.894427  0.447214
 -0.447214  0.894427
P factor:
2×2 Matrix{Float64}:
 0.894427  0.447214
 0.447214  1.34164

julia> isapprox(F.O * F.P, S)
true

julia> O, P = F; # destructuring via iteration

julia> O == F.O && P == F.P
true

SymplecticMatrices.Symplectic — Method

Symplectic(J::SymplecticForm, S::AbstractMatrix) <: AbstractMatrix

Construct a wrapper of a symplectic matrix S with its corresponding symplectic basis defined by the symplectic form J.

SymplecticMatrices.Williamson — Type

Williamson <: Factorization

Matrix factorization type of the williamson decomposition of a positive-definite matrix V. This is the return type of williamson(_), the corresponding matrix factorization function.

If F::Williamson is the factorization object, S and spectrum can be obtained via F.S and F.spectrum.

Iterating the decomposition produces the components S and spectrum.

Examples

julia> V = [7. 2.; 2. 1.]
2×2 Matrix{Float64}:
 7.0  2.0
 2.0  1.0

julia> isposdef(V)
true

julia> F = williamson(BlockForm(1), V)
Williamson{Float64, Matrix{Float64}, Vector{Float64}}
S factor:
2×2 Matrix{Float64}:
 0.448828  -1.95959
 0.61311   -0.448828
symplectic spectrum:
1-element Vector{Float64}:
 1.7320508075688772

julia> isapprox(F.S * V * F.S', Diagonal(repeat(F.spectrum, 2)))
true

julia> S, spectrum = F; # destructuring via iteration

julia> S == F.S && spectrum == F.spectrum
true

SymplecticMatrices.blochmessiah — Method

blochmessiah(form::SymplecticForm, S::AbstractMatrix) -> BlochMessiah
blochmessiah(S::Symplectic) -> BlochMessiah

Compute the Bloch-Messiah/Euler decomposition of a symplectic matrix S and return a BlockMessiah object.

The orthogonal symplectic matrices O and Q as well as the singular values values can be obtained via F.O, F.Q, and F.values, respectively.

Iterating the decomposition produces the components O, values, and Q, in that order.

Examples

julia> S = Symplectic(BlockForm(1), [1. 1.; 0. 1.])
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
 1.0  1.0
 0.0  1.0

julia> issymplectic(S)
true

julia> F = blochmessiah(S)
BlochMessiah{Float64, Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}, Vector{Float64}}
O factor:
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
 0.850651  -0.525731
 0.525731   0.850651
values:
1-element Vector{Float64}:
 1.618033988749895
Q factor:
2×2 Symplectic{BlockForm{Int64}, Float64, Matrix{Float64}}:
  0.525731  0.850651
 -0.850651  0.525731

julia> isapprox(S, F.O * Diagonal(vcat(F.values, F.values .^ (-1))) * F.Q, atol = 1e-10)
true

julia> issymplectic(F.O, atol = 1e-10) && issymplectic(F.Q, atol = 1e-10)
true

julia> O, values, Q = F; # destructuring via iteration

julia> O == F.O && values == F.values && Q == F.Q
true

SymplecticMatrices.issymplectic — Method

issymplectic(form::Symplecticform, x::AbstractMatrix; atol=0.0, rtol=atol)
issymplectic(x::Symplectic; atol=0.0, rtol=atol)

Return whether or not an input matrix is symplectic. Keyword arguments atol and rtol can be called to determine absolute and relative tolerances of the check, respectively.

SymplecticMatrices.polar — Method

polar(S::AbstractMatrix) -> Polar
polar(S::Symplectic) -> Polar

Compute the polar decomposition of a symplectic matrix S and return a Polar object.

O and P can be obtained from the factorization F via F.O and F.P, such that S = O * P. For the symplectic polar decomposition case, O is an orthogonal symplectic matrix and P is a positive-definite symmetric symplectic matrix.

Iterating the decomposition produces the components O and P.

Examples

julia> S = [1. 1.; 0. 1.]
2×2 Matrix{Float64}:
 1.0  1.0
 0.0  1.0

julia> issymplectic(BlockForm(1), S)
true

julia> F = polar(S)
Polar{Float64, Matrix{Float64}, Matrix{Float64}}
O factor:
2×2 Matrix{Float64}:
  0.894427  0.447214
 -0.447214  0.894427
P factor:
2×2 Matrix{Float64}:
 0.894427  0.447214
 0.447214  1.34164

julia> isapprox(F.O * F.P, S)
true

julia> O, P = F; # destructuring via iteration

julia> O == F.O && P == F.P
true

SymplecticMatrices.randsymplectic — Method

randsymplectic(form::SymplecticForm)
randsymplectic(::Symplectic, form::SymplecticForm)

Calculate a random symplectic matrix in symplectic representation defined by basis.

SymplecticMatrices.williamson — Method

williamson(form::SymplecticForm, V::AbstractMatrix) -> Williamson
williamson(::Symplectic, form::SymplecticForm, V::AbstractMatrix) -> Williamson

Compute the williamson decomposition of a positive-definite matrix V and return a Williamson object.

A symplectic matrix S and symplectic spectrum spectrum can be obtained via F.S and F.spectrum.

Iterating the decomposition produces the components S and spectrum.

Examples

julia> V = [7. 2.; 2. 1.]
2×2 Matrix{Float64}:
 7.0  2.0
 2.0  1.0

julia> isposdef(V)
true

julia> F = williamson(BlockForm(1), V)
Williamson{Float64, Matrix{Float64}, Vector{Float64}}
S factor:
2×2 Matrix{Float64}:
 0.448828  -1.95959
 0.61311   -0.448828
symplectic spectrum:
1-element Vector{Float64}:
 1.7320508075688772

julia> isapprox(F.S * V * F.S', Diagonal(repeat(F.spectrum, 2)))
true

julia> S, spectrum = F; # destructuring via iteration

julia> S == F.S && spectrum == F.spectrum
true